requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.11
aiosignal==1.3.2
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
discord.py==2.5.2
dotenv==0.9.9
frozenlist==1.6.2
idna==3.10
MemePy==1.2.2
multidict==6.4.4
Pillow==8.4.0
propcache==0.3.1
python-dotenv==1.1.0
requests==2.32.3
urllib3==2.4.0
yarl==1.20.0
